Creacion de la clase backpropagation

+++ loss_function.cpp
@@ -25,7 +25,7 @@ Loss::Loss(int epoch, const vector<int>& labels, const vector<vector<float>>& pr
 
 
 
-vector<vector<int>> distributionVector(const vector<int>& labels){
+vector<vector<float>> distributionVector(const vector<int>& labels){
     // La función tiene que leer el label y en un vector de 0 de 10 posiciones
     // pondrá un 1 en la posición que el label marque
 
@@ -33,7 +33,7 @@ vector<vector<int>> distributionVector(const vector<int>& labels){
 
     int cols = 10;
     auto rows = labels.size();
-    vector<vector<int>> positionVector(rows, vector<int>(cols, 0));
+    vector<vector<float>> positionVector(rows, vector<float>(cols, 0.0f));
     
     for (int i = 0; i < rows; i++) {
         int label = labels[i];
@@ -45,7 +45,7 @@ vector<vector<int>> distributionVector(const vector<int>& labels){
     return positionVector;
 }
 
-double cross_entropy(const vector<vector<int>>& distributionVector, vector<vector<float>> prediction) {
+double cross_entropy(const vector<vector<float>>& distributionVector, vector<vector<float>> prediction) {
     // Se ingresa el distributionVector, este vector fue implementado en la anterior
     // función, que como se explicó, es un vector que determina la posición
     // real del label
